+++ src/lib/index.js
@@ -50,7 +50,7 @@ let morning = {
         white : 'wh'
     },
     isMorning : true,
-    version : '0.12.39',
+    version : '0.12.40',
     map : {}
 };